+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased Changes for 0.6.x
 * Added `--watch` argument to `rojo build`. ([#284](https://github.com/rojo-rbx/rojo/pull/284))
 * Added dark theme support to plugin. ([#241](https://github.com/rojo-rbx/rojo/issues/241))
+* Added a revamped `rojo init` command, which will now create more complete projects.
 * Added the `rojo doc` command, which opens Rojo's documentation in your browser.
 * Simplified filesystem access code dramatically.
 * Improved error reporting and logging across the board.